requirements.txt
ultralytics
streamlit
numpy
pandas
matplotlib
torch
Pillow
opencv-python
opencv-python-headless>=4.5.0
googledrivedownloader
